refactor(core): extract PivotTable IO/filter/freq/sort/stats into modules
Decompose the monolithic PivotTable by delegating to focused helper
modules: pivot_io (sqlite/h5/anndata), pivot_filtering (by_freq/variance/
statistical_test), pivot_frequency (calculate/add_freq), pivot_sorting
(features/by_mutations/by_group), pivot_stats (mutation_enrichment_test).
Public methods keep their signatures and forward to the helpers.

Full test suite green (189 passed, 7 skipped).
